Include framemove.el
This is useful on Microsoft Windows, where the window manager is
proprietary and not particularly flexible. Maximizing a single
window (such as an Emacs frame) across multiple monitors can be a
frustrating ordeal, and even when manually done, the window is often
abruptly reset back to a single monitor's width.

Nvidia Mosaic (and whatever AMD calls theirs) can mitigate this, but
Mosaic is not supported on (at least some) Optimus laptops.

With this package, you can create a single (fullscreen) frame for each
monitor, and navigate as though they were actually a single combined
frame. It's pretty seamless.

It's not needed for Linux machines, where (at least under X.org)
multiple monitors can be joined into a single desktop irrespective of
the GPU driver.

Finally, just copy/pasting it in, because it never made it into
MELPA. Originally found it on the EmacsWiki.