Fix delete-window advice to account for tab-bar-mode
By default, attempting to close the last window in a frame will fail
and do nothing. I added some advice on emacswiki which, instead, would
delete the frame.

This causes issues when tab-bar-mode is used, because it will close
the frame, even when other tabs are open. Update the advice to account
for tab-bar-mode.

If tab-bar-mode is active, attempting to delete the last window will
instead call tab-bar-close-tab. This will close the tab. If it is also
the final tab, then tab-bar-mode will close the frame on its own.