Side-step completions when saving files
When saving a buffer to a new file, both ido and ivy (and probably
any autocomplete framework) try a little too hard to find existing
directories/folders. This commit adds a new function which temporarily
disables Ivy's directory name completion (on forward slash), and then
calls write-file interactively.

It also binds ivy-immediate-done to S-<return>, which is easier for me
to remember than C-M-j. This is needed to avoid overwriting similarly
named files.