Add a fallback functionality for hyper key
Getting the hyper modifier onto capslock is often pretty difficult, on
both Windows and Linux systems. Add some elisp that will map <menu> to
a one-off hyper leader key. This is useful because it's generally much
easier to map CAPS to <menu> in both Linux and Windows than to turn it
into a modifier.