Cache org-agenda-files intelligently
Maintain the value of org-agenda-files across sessions by saving and
restoring it from an *.el file. Add a save-hook to org-mode that adds
or removes the current file's path from the list depending on whether
or not the file has open tasks.

This replaces logic which just used every single file in the org
directory, which unfortunately takes forever.